+++ src/Commands/Base/PnPGraphCmdlet.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Graph;
 using Microsoft.SharePoint.Client;
+using PnP.Core.Services;
 using PnP.PowerShell.Commands.Attributes;
 using PnP.PowerShell.Commands.Model;
 using PnP.PowerShell.Commands.Properties;
@@ -20,6 +21,9 @@ namespace PnP.PowerShell.Commands.Base
         /// Reference the the SharePoint context on the current connection. If NULL it means there is no SharePoint context available on the current connection.
         /// </summary>
         public ClientContext ClientContext => Connection?.Context ?? PnPConnection.CurrentConnection.Context;
+
+        public PnPContext PnPContext => Connection?.PnPContext ?? PnPConnection.CurrentConnection.PnPContext;
+
 
         // do not remove '#!#99'
         [Parameter(Mandatory = false, HelpMessage = "Optional connection to be used by the cmdlet. Retrieve the value for this parameter by either specifying -ReturnConnection on Connect-PnPOnline or by executing Get-PnPConnection.")]
@@ -51,7 +55,7 @@ namespace PnP.PowerShell.Commands.Base
             {
                 if (PnPConnection.CurrentConnection?.ConnectionMethod == ConnectionMethod.ManagedIdentity)
                 {
-                    return TokenHandler.GetManagedIdentityTokenAsync(this,HttpClient, "https://graph.microsoft.com/").GetAwaiter().GetResult();
+                    return TokenHandler.GetManagedIdentityTokenAsync(this, HttpClient, "https://graph.microsoft.com/").GetAwaiter().GetResult();
                 }
                 else
                 {